Fix compile time Absinthe.Plug race condition
The if Code.ensure_loaded?(Absinthe.Plug) line removed by this commit
is evaluated at compile time, not run time. This means if
absinthe_plug is compiled later than this library the traced_pipeline
method will not be present.

It could work locally due to absinthe_plug being added and compiled
before this module. However if there is a different compilation order
in a CI build the function is not added causing an
UndefinedFunctionError failure.